FIX: float conversion removed

+++ mapie/estimator/classifier.py
@@ -216,7 +216,7 @@ class EnsembleClassifier(EnsembleEstimator):
             err_msg="The sum of the scores is not equal to one.",
             rtol=1e-5
         )
-        return y_pred_proba.astype(np.float64)
+        return y_pred_proba
 
     def _predict_proba_oof_estimator(
         self,
